Installer: experimental viewStepThree
